pyproject.toml
[tool.poetry]
name = "shield-ai"
version = "2.0.0"
description = "Intelligent Inventory Management System with AI-powered shrinkage prediction"
authors = ["Denis Egorov <denis.egorov.2017@gmail.com>"]
readme = "README.md"
packages = [{include = "shield_ai", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0"
alembic = "^1.12"
pandas = "^2.0"
openpyxl = "^3.1"
scipy = "^1.11"
streamlit = "^1.28"
plotly = "^5.17"
python-dotenv = "^1.0"
pydantic = "^2.5"
pydantic-settings = "^2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
black = "^24.3.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.0.0"
pre-commit = "^3.5.0"
bandit = "^1.7.0"
safety = "^2.3.0"
pandas-stubs = "^2.3.2.250926"
scipy-stubs = "^1.16.3.0"
pyright = "^1.1.300"
pylint = "^3.0"
pip-audit = "^2.9.0"
radon = "^6.0.1"
sphinx = "^8.2.3"

[tool.black]
line-length = 88
target-version = ["py312"]
include = "\\.(py|pyi)$"

[tool.isort]
profile = "black"
line_length = 10

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E402", "I"]

# Игнорировать E402 (import not at top of file) для файлов, где нужно изменить sys.path
[tool.ruff.lint.per-file-ignores]
"src/shield_ai/main_calibrate.py" = ["E402"]
"tests/integration/test_calibration.py" = ["E402"]
"tests/validation/test_data.py" = ["E402"]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src", "."]

[tool.coverage.run]
source = ["src/"]
omit = ["*/tests/*", "*/test_*.py", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__", "src/experimental", "src/typestubs"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
pythonPlatform = "Linux"
typeCheckingMode = "basic"

[tool.pip-audit]
require-hashes = false
strict = true

[tool.radon]
exclude = ["tests/", "docs/", "setup.py", "scripts/"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
